# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "YAOpenEvidence API",
    "description": "YAOpenEvidence 医学文献证据问答 API。\n\n- 问答任务异步执行：`POST /v1/answers` 返回 202，进度走 `GET /v1/jobs/{id}/events`（SSE）\n- 鉴权：管理员创建账号，登录后使用 `Authorization: Bearer <access_token>`（包括 SSE）\n- 错误：RFC 9457 Problem Details（`application/problem+json`），按 `code` 分支\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/users": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "创建账号",
        "operationId": "create_user_v1_users_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            },
            "headers": {
              "Location": {
                "schema": {
                  "type": "string"
                },
                "description": "新建用户的资源地址"
              }
            }
          },
          "default": {
            "description": "Problem Details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "users"
        ],
        "summary": "浏览账号",
        "operationId": "list_users_v1_users_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_UserRead_"
                }
              }
            }
          },
          "default": {
            "description": "Problem Details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/users/{user_id}": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "读取账号",
        "operationId": "get_user_v1_users__user_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            }
          },
          "default": {
            "description": "Problem Details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "users"
        ],
        "summary": "启用或禁用账号",
        "operationId": "update_user_v1_users__user_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            }
          },
          "default": {
            "description": "Problem Details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/users/{user_id}/password": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "重置密码并注销全部会话",
        "operationId": "reset_password_v1_users__user_id__password_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "default": {
            "description": "Problem Details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/answers/{answer_id}/papers/{n}": {
      "get": {
        "tags": [
          "answers"
        ],
        "summary": "本次阅读的第 n 篇（笔记、引文、段落、全文）",
        "operationId": "get_answer_paper_v1_answers__answer_id__papers__n__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "answer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Answer Id"
            }
          },
          {
            "name": "n",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "title": "N"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnswerPaperDetail"
                }
              }
            }
          },
          "default": {
            "description": "Problem Details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/papers": {
      "get": {
        "tags": [
          "papers"
        ],
        "summary": "浏览文献库",
        "operationId": "list_papers_v1_papers_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Q"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_PaperMeta_"
                }
              }
            }
          },
          "default": {
            "description": "Problem Details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/papers/{key}": {
      "get": {
        "tags": [
          "papers"
        ],
        "summary": "读取文献元数据",
        "operationId": "get_paper_v1_papers__key__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaperMeta"
                }
              }
            }
          },
          "default": {
            "description": "Problem Details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/papers/{key}/paragraphs": {
      "get": {
        "tags": [
          "papers"
        ],
        "summary": "读取全部段落",
        "operationId": "get_paragraphs_v1_papers__key__paragraphs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParagraphList"
                }
              }
            }
          },
          "default": {
            "description": "Problem Details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/papers/{key}/paragraphs/{pid}": {
      "get": {
        "tags": [
          "papers"
        ],
        "summary": "读取单个段落",
        "operationId": "get_paragraph_v1_papers__key__paragraphs__pid__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key"
            }
          },
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Pid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Paragraph"
                }
              }
            }
          },
          "default": {
            "description": "Problem Details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/papers/{key}/facts": {
      "get": {
        "tags": [
          "papers"
        ],
        "summary": "读取原子事实",
        "operationId": "get_facts_v1_papers__key__facts_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FactList"
                }
              }
            }
          },
          "default": {
            "description": "Problem Details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/papers/{key}/fulltext": {
      "get": {
        "tags": [
          "papers"
        ],
        "summary": "读取带段落锚点的全文",
        "operationId": "get_fulltext_v1_papers__key__fulltext_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/markdown": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "description": "Problem Details (RFC 9457)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Request validation failed",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnswerPaperDetail": {
        "properties": {
          "n": {
            "type": "integer",
            "title": "N"
          },
          "pmid": {
            "type": "string",
            "title": "Pmid",
            "default": ""
          },
          "doi": {
            "type": "string",
            "title": "Doi",
            "default": ""
          },
          "pmcid": {
            "type": "string",
            "title": "Pmcid",
            "default": ""
          },
          "title": {
            "type": "string",
            "title": "Title",
            "default": ""
          },
          "year": {
            "type": "string",
            "title": "Year",
            "default": ""
          },
          "journal": {
            "type": "string",
            "title": "Journal",
            "default": ""
          },
          "issn": {
            "type": "string",
            "title": "Issn",
            "default": ""
          },
          "authors": {
            "type": "string",
            "title": "Authors",
            "default": ""
          },
          "quartile": {
            "type": "string",
            "title": "Quartile",
            "default": ""
          },
          "rank_label": {
            "type": "string",
            "title": "Rank Label",
            "default": ""
          },
          "source": {
            "type": "string",
            "enum": [
              "pmc",
              "pdf",
              "inst",
              "abstract"
            ],
            "title": "Source",
            "default": "abstract"
          },
          "relevance": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relevance"
          },
          "n_paragraphs": {
            "type": "integer",
            "title": "N Paragraphs",
            "default": 0
          },
          "n_citations": {
            "type": "integer",
            "title": "N Citations",
            "default": 0
          },
          "n_citations_verified": {
            "type": "integer",
            "title": "N Citations Verified",
            "default": 0
          },
          "notes_md": {
            "type": "string",
            "title": "Notes Md",
            "default": ""
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/VerifiedQuote"
            },
            "type": "array",
            "title": "Citations",
            "default": []
          },
          "facts": {
            "items": {
              "$ref": "#/components/schemas/Fact"
            },
            "type": "array",
            "title": "Facts",
            "default": []
          },
          "paragraphs": {
            "items": {
              "$ref": "#/components/schemas/Paragraph"
            },
            "type": "array",
            "title": "Paragraphs",
            "default": []
          },
          "fulltext_md": {
            "type": "string",
            "title": "Fulltext Md",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "n"
        ],
        "title": "AnswerPaperDetail"
      },
...
      "CreateUserRequest": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 64,
            "minLength": 3,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_.-]{2,63}$",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 12,
            "format": "password",
            "title": "Password",
            "writeOnly": true
          },
          "role": {
            "type": "string",
            "enum": [
              "user",
              "admin"
            ],
            "title": "Role",
            "default": "user"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "CreateUserRequest"
      },
...
      "FactList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Fact"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "FactList"
      },
...
      "Page_PaperMeta_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PaperMeta"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[PaperMeta]"
      },
      "Page_UserRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/UserRead"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[UserRead]"
      },
      "PaperMeta": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "pmid": {
            "type": "string",
            "title": "Pmid",
            "default": ""
          },
          "doi": {
            "type": "string",
            "title": "Doi",
            "default": ""
          },
          "pmcid": {
            "type": "string",
            "title": "Pmcid",
            "default": ""
          },
          "title": {
            "type": "string",
            "title": "Title",
            "default": ""
          },
          "year": {
            "type": "string",
            "title": "Year",
            "default": ""
          },
          "journal": {
            "type": "string",
            "title": "Journal",
            "default": ""
          },
          "issn": {
            "type": "string",
            "title": "Issn",
            "default": ""
          },
          "quartile": {
            "type": "string",
            "title": "Quartile",
            "default": ""
          },
          "authors": {
            "type": "string",
            "title": "Authors",
            "default": ""
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": ""
          },
          "types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Types",
            "default": []
          },
          "indexed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Indexed At"
          },
          "n_paragraphs": {
            "type": "integer",
            "title": "N Paragraphs",
            "default": 0
          },
          "n_facts": {
            "type": "integer",
            "title": "N Facts",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "key"
        ],
        "title": "PaperMeta"
      },
...
      "ParagraphList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Paragraph"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "ParagraphList"
      },
...
      "PasswordResetRequest": {
        "properties": {
          "new_password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 12,
            "format": "password",
            "title": "New Password",
            "writeOnly": true
          }
        },
        "type": "object",
        "required": [
          "new_password"
        ],
        "title": "PasswordResetRequest"
      },
...
      "UserRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "role": {
            "type": "string",
            "enum": [
              "user",
              "admin"
            ],
            "title": "Role"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "username",
          "role",
          "is_active",
          "created_at"
        ],
        "title": "UserRead"
      },
      "UserUpdateRequest": {
        "properties": {
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "is_active"
        ],
        "title": "UserUpdateRequest"
      },
...
      "Problem": {
        "description": "RFC 9457 Problem Details。`code` 是稳定枚举，`detail` 仅供人读。",
        "properties": {
          "type": {
            "examples": [
              "urn:yaoe:error:not_found"
            ],
            "title": "Type",
            "type": "string"
          },
          "title": {
            "examples": [
              "Not Found"
            ],
            "title": "Title",
            "type": "string"
          },
          "status": {
            "examples": [
              404
            ],
            "title": "Status",
            "type": "integer"
          },
          "detail": {
            "examples": [
              "answer 'abc' not found"
            ],
            "title": "Detail",
            "type": "string"
          },
          "instance": {
            "examples": [
              "/v1/answers/abc"
            ],
            "title": "Instance",
            "type": "string"
          },
          "code": {
            "examples": [
              "not_found"
            ],
            "title": "Code",
            "type": "string"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Errors"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "code"
        ],
        "title": "Problem",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "登录返回的可撤销会话令牌",
        "scheme": "bearer"
      }
    }
  }
}